uHAL: Use STL function and bind instead of boost. See #226
